C cleanup: be more explicit about conversions involvig `float` and `double`
byterun/bigarray.c, byterun/ints.c:
  add explicit casts to clarify the intent
  renamed float field of conversion union from `d` to `f`.

byterun/compact.c, byterun/gc_ctrl.c:
  some local variables were of type `float` while all FP computations
  here are done in double precision;
  turned these variables into `double`.

[-Wdouble-promotion -Wfloat-conversion]

+++ byterun/bigarray.c
@@ -573,7 +573,7 @@ value caml_ba_get_N(value vb, value * vind, int nind)
   default:
     CAMLassert(0);
   case CAML_BA_FLOAT32:
-    return caml_copy_double(((float *) b->data)[offset]);
+    return caml_copy_double((double) ((float *) b->data)[offset]);
   case CAML_BA_FLOAT64:
     return caml_copy_double(((double *) b->data)[offset]);
   case CAML_BA_SINT8:
@@ -594,7 +594,7 @@ value caml_ba_get_N(value vb, value * vind, int nind)
     return Val_long(((intnat *) b->data)[offset]);
   case CAML_BA_COMPLEX32:
     { float * p = ((float *) b->data) + offset * 2;
-      return copy_two_doubles(p[0], p[1]); }
+      return copy_two_doubles((double) p[0], (double) p[1]); }
   case CAML_BA_COMPLEX64:
     { double * p = ((double *) b->data) + offset * 2;
       return copy_two_doubles(p[0], p[1]); }
